+++ src/gocept/selenium/scripts/tests/test_converthtmltests.py
@@ -52,7 +52,7 @@ class TestConversion(unittest.TestCase):
         parser = make_parser()
         parser.error = mock.Mock()
         parse_options(parser, [])
-        self.failUnless(parser.error.called)
+        self.assertTrue(parser.error.called)
         parser.error.assert_called_with(LAYER_REQUIRED)
 
     def test_parse_options_no_directory(self):
@@ -62,7 +62,7 @@ class TestConversion(unittest.TestCase):
         parser = make_parser()
         parser.error = mock.Mock()
         parse_options(parser, ['-l', LAYER])
-        self.failUnless(parser.error.called)
+        self.assertTrue(parser.error.called)
         parser.error.assert_called_with(DIRECTORY_REQUIRED)
 
     def test_parse_options_no_module(self):
@@ -72,7 +72,7 @@ class TestConversion(unittest.TestCase):
         parser = make_parser()
         parser.error = mock.Mock()
         parse_options(parser, ['-l', 'layer', 'dummy'])
-        self.failUnless(parser.error.called)
+        self.assertTrue(parser.error.called)
         parser.error.assert_called_with(LAYER_WITH_MODULE)
 
     def test_parse_options_one_directory(self):
@@ -82,7 +82,7 @@ class TestConversion(unittest.TestCase):
         parser = make_parser()
         parser.error = mock.Mock()
         parse_options(parser, ['-l', LAYER, 'first', 'second'])
-        self.failUnless(parser.error.called)
+        self.assertTrue(parser.error.called)
         parser.error.assert_called_with(ONE_DIRECTORY)
 
     def test_parse_options_directory_not_exist(self):
@@ -93,7 +93,7 @@ class TestConversion(unittest.TestCase):
         parser = make_parser()
         parser.error = mock.Mock()
         parse_options(parser, ['-l', LAYER, 'first'])
-        self.failUnless(parser.error.called)
+        self.assertTrue(parser.error.called)
         parser.error.assert_called_with(DIRECTORY_NOT_EXIST)
 
     def test_parse_options_directory_exists(self):
@@ -104,10 +104,10 @@ class TestConversion(unittest.TestCase):
         parser.error = mock.Mock()
         source = tempfile.gettempdir()
         options, directory = parse_options(parser, ['-l', LAYER, source])
-        self.assertEquals(source, directory)
-        self.assertEquals(options.layer, LAYER)
-        self.assertEquals(options.target, DEFAULT_TARGET)
-        self.failUnless(options.verbose)
+        self.assertEqual(source, directory)
+        self.assertEqual(options.layer, LAYER)
+        self.assertEqual(options.target, DEFAULT_TARGET)
+        self.assertTrue(options.verbose)
 
     def test_parse_options_quiet(self):
         from gocept.selenium.scripts.converthtmltests import parse_options
@@ -116,7 +116,7 @@ class TestConversion(unittest.TestCase):
         parser.error = mock.Mock()
         source = tempfile.mkdtemp()
         options, directory = parse_options(parser, ['-q', '-l', LAYER, source])
-        self.failIf(options.verbose)
+        self.assertFalse(options.verbose)
 
     def test_parse_options_target(self):
         from gocept.selenium.scripts.converthtmltests import parse_options
@@ -127,7 +127,7 @@ class TestConversion(unittest.TestCase):
         target = tempfile.mktemp()
         options, directory = parse_options(
             parser, ['-f', target, '-l', LAYER, source])
-        self.assertEquals(options.target, target)
+        self.assertEqual(options.target, target)
 
     def test_parse_file(self):
         from gocept.selenium.scripts.converthtmltests import parse_file
@@ -137,11 +137,11 @@ class TestConversion(unittest.TestCase):
         filename = os.path.join(tests_dir, 'plone3login.html')
 
         testname, commands, encoding = parse_file(filename)
-        self.assertEquals(encoding, 'utf-8')
-        self.assertEquals(testname, 'plone3login')
-        self.assertEquals(len(commands), 5)
-        self.assertEquals('        selenium.open("/plone/login_form")',
-                          commands[0])
+        self.assertEqual(encoding, 'utf-8')
+        self.assertEqual(testname, 'plone3login')
+        self.assertEqual(len(commands), 5)
+        self.assertEqual('        selenium.open("/plone/login_form")',
+                         commands[0])
 
     def test_parse_file_no_title(self):
         from gocept.selenium.scripts.converthtmltests import parse_file
@@ -151,7 +151,7 @@ class TestConversion(unittest.TestCase):
         filename = os.path.join(tests_dir, 'notitle.html')
 
         testname, commands, encoding = parse_file(filename)
-        self.assertEquals(None, testname)
+        self.assertEqual(None, testname)
 
     def test_parse_file_no_commands(self):
         from gocept.selenium.scripts.converthtmltests import parse_file
@@ -161,33 +161,33 @@ class TestConversion(unittest.TestCase):
         filename = os.path.join(tests_dir, 'nocommands.html')
 
         testname, commands, encoding = parse_file(filename)
-        self.assertEquals('nocommands', testname)
-        self.assertEquals(len(commands), 0)
+        self.assertEqual('nocommands', testname)
+        self.assertEqual(len(commands), 0)
 
     def test_formatcommand_no_command(self):
         from gocept.selenium.scripts.converthtmltests import formatcommand
         line = formatcommand('')
-        self.assertEquals(line, '')
+        self.assertEqual(line, '')
 
     def test_formatcommand_command_without_args(self):
         from gocept.selenium.scripts.converthtmltests import formatcommand
         line = formatcommand('command')
-        self.assertEquals(line, '        selenium.command()')
+        self.assertEqual(line, '        selenium.command()')
 
     def test_formatcommand_command_with_single_arg(self):
         from gocept.selenium.scripts.converthtmltests import formatcommand
         line = formatcommand('command', 'arg')
-        self.assertEquals(line, '        selenium.command("arg")')
+        self.assertEqual(line, '        selenium.command("arg")')
 
     def test_formatcommand_command_with_two_args(self):
         from gocept.selenium.scripts.converthtmltests import formatcommand
         line = formatcommand('command', 'arg1', 'arg2')
-        self.assertEquals(line, '        selenium.command("arg1", "arg2")')
+        self.assertEqual(line, '        selenium.command("arg1", "arg2")')
 
     def test_formatcommand_command_with_three_args_one_empty(self):
         from gocept.selenium.scripts.converthtmltests import formatcommand
         line = formatcommand('command', 'arg1', '', 'arg2')
-        self.assertEquals(line, '        selenium.command("arg1", "arg2")')
+        self.assertEqual(line, '        selenium.command("arg1", "arg2")')
 
     def test_parse_directory(self):
         from gocept.selenium.scripts.converthtmltests import parse_directory
@@ -197,8 +197,8 @@ class TestConversion(unittest.TestCase):
 
         tests = [test for test in parse_directory(tests_dir, False)]
         tests.pop(0)  # this is the encoding
-        self.assertEquals(len(tests), 1)
-        self.assertEquals(tests[0], PLONE3LOGIN_METHOD)
+        self.assertEqual(len(tests), 1)
+        self.assertEqual(tests[0], PLONE3LOGIN_METHOD)
 
     def test_parse_directory_no_html(self):
         from gocept.selenium.scripts.converthtmltests import parse_directory
@@ -207,13 +207,13 @@ class TestConversion(unittest.TestCase):
         tests_dir = os.path.dirname(gocept.selenium.scripts.__file__)
 
         tests = [test for test in parse_directory(tests_dir, False)]
-        self.assertEquals(len(tests), 0)
+        self.assertEqual(len(tests), 0)
 
     def test_make_module(self):
         from gocept.selenium.scripts.converthtmltests import make_module
 
         module = make_module([PLONE3LOGIN_METHOD], LAYER, 'module', 'utf-8')
-        self.assertEquals(module, PLONE3LOGIN_MODULE)
+        self.assertEqual(module, PLONE3LOGIN_MODULE)
 
     def test_main(self):
         from gocept.selenium.scripts.converthtmltests import main
@@ -227,18 +227,18 @@ class TestConversion(unittest.TestCase):
         try:
             main(['-f', target, '-l', LAYER, tests_dir])
             module = open(target).read()
-            self.assertEquals(module, PLONE3LOGIN_MODULE)
+            self.assertEqual(module, PLONE3LOGIN_MODULE)
             text = output.getvalue()
             lines = text.splitlines()
-            self.assertEquals(len(lines), 4)
-            self.failUnless(lines[0].startswith('Parsing ['))
-            self.failUnless('plone3login.html]' in text)
-            self.failUnless(lines[1].startswith('Parsing ['))
-            self.failUnless('nocommands.html]' in text)
-            self.failUnless(lines[2].startswith('Parsing ['))
-            self.failUnless('notitle.html]' in text)
-            self.failUnless(lines[3].startswith('Generating ['))
-            self.failUnless(target in lines[3])
+            self.assertEqual(len(lines), 4)
+            self.assertTrue(lines[0].startswith('Parsing ['))
+            self.assertIn('plone3login.html]', text)
+            self.assertTrue(lines[1].startswith('Parsing ['))
+            self.assertIn('nocommands.html]', text)
+            self.assertTrue(lines[2].startswith('Parsing ['))
+            self.assertIn('notitle.html]', text)
+            self.assertTrue(lines[3].startswith('Generating ['))
+            self.assertIn(target, lines[3])
         finally:
             sys.stdout = sys.__stdout__
 
@@ -253,7 +253,7 @@ class TestConversion(unittest.TestCase):
         sys.stdout = output
         try:
             main(['-f', target, '-l', LAYER, tests_dir])
-            self.failIf(os.path.exists(target))
-            self.assertEquals(output.getvalue(), 'No file was generated !\n')
+            self.assertFalse(os.path.exists(target))
+            self.assertEqual(output.getvalue(), 'No file was generated !\n')
         finally:
             sys.stdout = sys.__stdout__