+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-4.0 (unreleased)
+4.0 (2018-11-09)
 ----------------
 
 - Drop support for Zope 2.